# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CompletionRequestBody": {
        "properties": {
          "max_tokens": {
            "default": 4096,
            "maximum": 128000.0,
            "minimum": 1.0,
            "title": "Max Tokens",
            "type": "integer"
          },
          "prompt": {
            "description": "The prompt to complete",
            "minLength": 1,
            "title": "Prompt",
            "type": "string"
          },
          "temperature": {
            "default": 0.7,
            "maximum": 2.0,
            "minimum": 0.0,
            "title": "Temperature",
            "type": "number"
          },
          "tier": {
            "description": "Model tier name (e.g. 'high_coding')",
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "tier",
          "prompt"
        ],
        "title": "CompletionRequestBody",
        "type": "object"
      },
      "CompletionResponseBody": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "endpoint_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Endpoint Id"
          },
          "outcome": {
            "title": "Outcome",
            "type": "string"
          },
          "served_from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Served From"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "CompletionResponseBody",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ForgeOps API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ai/complete": {
      "post": {
        "description": "Run a completion with the fixed security/admission order.\n\nOrder: OIDC \u2192 claims.sub \u2192 Redis limiter \u2192 cache \u2192 router/provider.",
        "operationId": "complete_api_v1_ai_complete_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompletionRequestBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompletionResponseBody"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete",
        "tags": [
          "ai"
        ]
      }
    },
...
  }
}